Avoid creating new String objects by instead modifying existing ones.
In Ruby, it is generally preferred to modify existing strings instead of
creating new strings. In Ruby, unlike many other languages, strings are
mutable. This is an imperative language, not a functional one, so let's
make use of it!

So, instead of using "str1 + str2", which creates a *new* string object,
we use "str1 << str2", which *modifies* str1 by appending the contents
of str2 to it.