Replace calls to constructors (Array.new, Hash.new) with literals.
It is generally preferred to use Hash and Array literals ({} and [])
instead of Hash.new and Array.new, unless one actually needs to pass
a parameter. Also, a lot of other collection classes provide a '[]'
class method as a synonym for 'new', so that you can e.g. call
Set[1, 2, 3] instead of Set.new(1, 2, 3). I prefer this style because
it is visually similar to an array literal and could be read as "cast
an array to a Set".