Don't initialize variables to nil (or false).
Unitialized variables evaluate to nil anyway, so initializing them to
nil isn't really necessary. Also, since nil is considered false in a
boolean context, uninitialized variables will evaluate to false and
need not be initialized to false, unless the exact value false is
required, e.g. for serialization purposes (in YAML, for example, nil
and false are treated differently).